Refactor grpc client function organization
